updated folder structure + changed functions to arrow functions
