+++ App.js
@@ -1,19 +1,17 @@
-
 import React from 'react';
-import Providers from './navigation'
+import Providers from './navigation';
 // import {NavigationContainer} from '@react-navigation/native';
 
 // import MyTabs from './components/NavigationBar/MyTabs';
 // import AuthStack from './navigation/AuthStack';
 
 const App = () => {
-  return  <Providers/>
-
+  return <Providers />;
 
   //   <NavigationContainer>
   //     <AuthStack />
   //     {/* <MyTabs/> */}
   //   </NavigationContainer>
   // )
-}
+};
 export default App;
