+++ navigation/AppStack.js
@@ -1,16 +1,46 @@
-import React from 'react';
-import {createStackNavigator } from '@react-navigation/stack'
-import HomeScreen from  '../screens/HomeScreen'
-import { StackActions } from '@react-navigation/native';
+import * as React from 'react';
+import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 
-const Stack = createStackNavigator();
+import EbookScreen from '../screens/EbookScreen';
+import SearchScreen from '../../screens/SearchScreen';
+import LibraryScreen from '../../screens/LibraryScreen';
+import ProfileScreen from '../../screens/ProfileScreen';
 
-const AppStack = () => { 
-    return ( 
-        <Stack.Navigator>
-            <Stack.Screen name='Home' component={HomeScreen}/>
-        </Stack.Navigator>
-    )
-}
+import Icon from 'react-native-vector-icons/MaterialIcons';
 
-export default AppStack
+const Tab = createBottomTabNavigator();
+
+const AppStack = () => {
+  return (
+    <Tab.Navigator
+      screenOptions={({route}) => ({
+        tabBarIcon: ({focused, color, size}) => {
+          let iconName;
+
+          if (route.name === 'Home') {
+            iconName = focused ? 'home' : 'home';
+          } else if (route.name === 'Search') {
+            iconName = focused ? 'search' : 'search';
+          } else if (route.name === 'Library') {
+            iconName = focused ? 'book' : 'book';
+          } else if (route.name === 'Profile') {
+            iconName = focused ? 'person' : 'person';
+          }
+
+          // You can return any component that you like here!
+          return <Icon name={iconName} size={size} color={color} />;
+        },
+      })}
+      tabBarOptions={{
+        activeTintColor: 'blue',
+        inactiveTintColor: 'gray',
+      }}>
+      <Tab.Screen name="Home" component={EbookScreen} />
+      <Tab.Screen name="Search" component={SearchScreen} />
+      <Tab.Screen name="Library" component={LibraryScreen} />
+      <Tab.Screen name="Profile" component={ProfileScreen} />
+    </Tab.Navigator>
+  );
+};
+
+export default AppStack;
