+++ screens/LibraryScreen.js
@@ -1,13 +1,13 @@
 import * as React from 'react';
 import {Text, StyleSheet, View} from 'react-native';
 
-function ProfileScreen({navigation}) {
+const LibraryScreen = () => {
   return (
     <View style={styles.container}>
-      <Text>Profile Screen</Text>
+      <Text>Library Screen</Text>
     </View>
   );
-}
+};
 
 const styles = StyleSheet.create({
   container: {
@@ -17,4 +17,4 @@ const styles = StyleSheet.create({
   },
 });
 
-export default ProfileScreen;
+export default LibraryScreen;
